Move migrate dialog into savings
