+++ packages/app/src/features/dialogs/savings/migrate/downgrade/DowngradeDialog.tsx
@@ -1,6 +1,6 @@
 import { Token } from '@/domain/types/Token'
+import { CommonDialogProps } from '@/features/dialogs/common/types'
 import { Dialog, DialogContent } from '@/ui/atoms/dialog/Dialog'
-import { CommonDialogProps } from '../../common/types'
 import { DowngradeDialogContentContainer } from './DowngradeDialogContentContainer'
 
 interface DowngradeDialogProps extends CommonDialogProps {
