+++ packages/app/src/features/dialogs/savings/migrate/downgrade/DowngradeDialogContentContainer.tsx
@@ -1,7 +1,7 @@
 import { Token } from '@/domain/types/Token'
+import { DialogContentSkeleton } from '@/features/dialogs/common/components/skeletons/DialogContentSkeleton'
+import { SuccessView } from '@/features/dialogs/common/views/SuccessView'
 import { withSuspense } from '@/ui/utils/withSuspense'
-import { DialogContentSkeleton } from '../../common/components/skeletons/DialogContentSkeleton'
-import { SuccessView } from '../../common/views/SuccessView'
 import { useMigrateDialog } from '../common/logic/useMigrateDialog'
 import { DowngradeUSDSToDaiView } from './views/DowngradeUSDSToDaiView'
