+++ packages/app/src/features/savings/logic/makeMigrationInfo.ts
@@ -3,8 +3,8 @@ import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { Percentage } from '@/domain/types/NumericValues'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { TokensInfo } from '@/domain/wallet/useTokens/TokenInfo'
-import { DowngradeDialog } from '@/features/dialogs/migrate/downgrade/DowngradeDialog'
-import { UpgradeDialog } from '@/features/dialogs/migrate/upgrade/UpgradeDialog'
+import { DowngradeDialog } from '@/features/dialogs/savings/migrate/downgrade/DowngradeDialog'
+import { UpgradeDialog } from '@/features/dialogs/savings/migrate/upgrade/UpgradeDialog'
 import { assert } from '@/utils/assert'
 import { determineApyImprovement } from './determineApyImprovement'
